get all in master

+++ src/app/master/brand-table/brand-table.component.ts
@@ -1,3 +1,4 @@
+import { DeleteConfirmationComponent } from './../../delete-confirmation/delete-confirmation.component';
 import { BrandModalComponent } from './../brand-modal/brand-modal.component';
 import { ApiService } from './../../api.service';
 import { AfterViewInit, Component, ViewChild } from '@angular/core';
@@ -53,10 +54,17 @@ export class BrandTableComponent implements AfterViewInit {
   }
 
   deleteBrand(id: any) {
-    this.apiService.deleteBrand(id).subscribe(response => {
-      this.loadBrand()
-      console.log('Delete successful', response);
-    })
+    const dialogRef = this.dialog.open(DeleteConfirmationComponent);
+
+    dialogRef.afterClosed().subscribe((result) => {
+      if (result) {
+        // User confirmed deletion
+        this.apiService.deleteBrand(id).subscribe(response => {
+          this.loadBrand();
+          console.log('Delete successful', response);
+        });
+      }
+    });
   }
 
   Openpopup(id: any, title: any) {
